+++ apps/prairielearn/src/ee/lib/billing/plans.ts
@@ -7,8 +7,8 @@ import {
   PlanGrant,
   PlanGrantSchema,
 } from '../../../lib/db-types';
-import { PLAN_NAMES, PlanName } from './plans-types';
-import { insertPlanGrant, updatePlanGrant, deletePlanGrant } from '../../models/plan-grants';
+import { PLAN_NAMES, PlanFeatureName, PlanName, getFeaturesForPlans } from './plans-types';
+import { ensurePlanGrant, updatePlanGrant, deletePlanGrant } from '../../models/plan-grants';
 import {
   insertCourseInstanceRequiredPlan,
   deleteCourseInstanceRequiredPlan,
@@ -197,7 +197,7 @@ async function reconcilePlanGrants(
   );
 
   for (const plan of newPlans) {
-    await insertPlanGrant({
+    await ensurePlanGrant({
       plan_grant: {
         ...context,
         plan_name: plan.plan,
@@ -235,4 +235,36 @@ async function getInstitutionForCourseInstance(course_instance_id: string): Prom
     { course_instance_id },
     InstitutionSchema,
   );
+}
+
+function getPlansForPlanGrants(planGrants: PlanGrant[]): PlanName[] {
+  const plans = new Set<PlanName>();
+  for (const planGrant of planGrants) {
+    plans.add(planGrant.plan_name);
+  }
+  return Array.from(plans);
+}
+
+/**
+ * Returns whether or not the given plan grants match all of the given features.
+ */
+export function planGrantsMatchFeatures(
+  planGrants: PlanGrant[],
+  features: PlanFeatureName[],
+): boolean {
+  const grantedPlans = getPlansForPlanGrants(planGrants);
+  const grantedFeatures = getFeaturesForPlans(grantedPlans);
+  return (
+    grantedFeatures.length === features.length &&
+    grantedFeatures.every((feature) => features.includes(feature))
+  );
+}
+
+/**
+ * Given a list of existing plan grants and a list of required plans, returns
+ * a list of plans that are required but not granted.
+ */
+export function getMissingPlanGrants(existingPlanGrants: PlanGrant[], requiredPlans: PlanName[]) {
+  const existingPlans = getPlansForPlanGrants(existingPlanGrants);
+  return requiredPlans.filter((plan) => !existingPlans.includes(plan));
 }