+++ apps/prairielearn/src/pages/administratorCourseRequests/administratorCourseRequests.js
@@ -76,17 +76,14 @@ router.post('/', (req, res, next) => {
         if (ERR(err, next)) return;
 
         res.redirect(`/pl/administrator/jobSequence/${job_id}/`);
-        opsbot.sendCourseRequestMessage(
-          `*Creating course*\n` +
-            `Course rubric: ${repo_options.short_name}\n` +
-            `Course title: ${repo_options.title}\n` +
-            `Approved by: ${res.locals.authn_user.name}`,
-          (err) => {
-            ERR(err, () => {
-              logger.error(err);
-            });
-          },
-        );
+        opsbot
+          .sendCourseRequestMessage(
+            `*Creating course*\n` +
+              `Course rubric: ${repo_options.short_name}\n` +
+              `Course title: ${repo_options.title}\n` +
+              `Approved by: ${res.locals.authn_user.name}`,
+          )
+          .catch((err) => logger.error(err));
       });
     });
   } else {