+++ apps/prairielearn/src/pages/administratorFeatures/administratorFeatures.html.ts
@@ -162,7 +162,7 @@ function FeatureGrantBreadcrumbs({ featureGrant }: { featureGrant: FeatureGrantR
       ${
         hasCourse
           ? html`<li class="list-inline-item inline-flex">
-              ${featureGrant.course_title} (${featureGrant.course_short_name})
+              ${featureGrant.course_short_name}: ${featureGrant.course_title}
             </li>`
           : null
       }
@@ -176,7 +176,7 @@ function FeatureGrantBreadcrumbs({ featureGrant }: { featureGrant: FeatureGrantR
       ${
         hasUser
           ? html`<li class="list-inline-item inline-flex">
-              ${featureGrant.user_name} (${featureGrant.user_uid})
+              ${featureGrant.user_uid} (${featureGrant.user_name})
             </li>`
           : null
       }
@@ -248,14 +248,14 @@ export function AddFeatureGrantModalBody({
           id="feature-grant-institution"
           name="institution_id"
         >
-          <option value="">None</option>
+          <option value="">All institutions</option>
           ${institutions.map((institution) => {
             return html`
               <option
                 value="${institution.id}"
                 ${institution.id === institution_id ? 'selected' : ''}
               >
-                ${institution.long_name}
+                ${institution.long_name} (${institution.short_name})
               </option>
             `;
           })}
@@ -273,11 +273,11 @@ export function AddFeatureGrantModalBody({
           name="course_id"
           ${!institution_id ? 'disabled' : ''}
         >
-          <option value="">None</option>
+          <option value="">All courses in this institution</option>
           ${(courses ?? []).map((course) => {
             return html`
               <option value="${course.id}" ${course.id === course_id ? 'selected' : ''}>
-                ${course.title}
+                ${course.short_name}: ${course.title}
               </option>
             `;
           })}
@@ -295,14 +295,14 @@ export function AddFeatureGrantModalBody({
           name="course_instance_id"
           ${!course_id ? 'disabled' : ''}
         >
-          <option value="">None</option>
+          <option value="">All courses instances in this course</option>
           ${(course_instances ?? []).map((course_instance) => {
             return html`
               <option
                 value="${course_instance.id}"
                 ${course_instance.id === course_instance_id ? 'selected' : ''}
               >
-                ${course_instance.long_name}
+                ${course_instance.long_name} (${course_instance.short_name})
               </option>
             `;
           })}