+++ apps/prairielearn/src/tests/permissions/effectiveUser.test.js
@@ -1,4 +1,3 @@
-const util = require('util');
 const assert = require('chai').assert;
 const { step } = require('mocha-steps');
 const { config } = require('../../lib/config');
@@ -6,7 +5,7 @@ const sqldb = require('@prairielearn/postgres');
 const sql = sqldb.loadSqlEquiv(__filename);
 const helperServer = require('../helperServer');
 const helperClient = require('../helperClient');
-const { insertEnrollment } = require('../../models/enrollment');
+const { ensureEnrollment } = require('../../models/enrollment');
 
 describe('effective user', function () {
   this.timeout(60000);
@@ -19,9 +18,7 @@ describe('effective user', function () {
   context.pageUrlStudent = `${context.baseUrl}/course_instance/1`;
   context.userId = 2;
 
-  before('set up testing server', async function () {
-    await util.promisify(helperServer.before().bind(this))();
-  });
+  before('set up testing server', helperServer.before().bind(this));
 
   before('insert users', async function () {
     await sqldb.callAsync('users_select_or_insert', [
@@ -54,7 +51,7 @@ describe('effective user', function () {
       'Editor',
       2,
     ]);
-    await insertEnrollment({
+    await ensureEnrollment({
       user_id: 4,
       course_instance_id: 1,
     });