+++ packages/config/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @prairielearn/config
 
+## 2.0.7
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+- Updated dependencies [f03853d90]
+  - @prairielearn/aws-imds@1.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- 6cad75197: Upgrade all dependencies
+- Updated dependencies [6cad75197]
+  - @prairielearn/aws-imds@1.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- 10cc07dcc: Upgrade all dependencies
+- Updated dependencies [10cc07dcc]
+  - @prairielearn/aws-imds@1.0.5
+
 ## 2.0.4
 
 ### Patch Changes