+++ packages/html-ejs/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @prairielearn/html-ejs
 
+## 1.1.9
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+- Updated dependencies [f03853d90]
+  - @prairielearn/html@3.0.7
+
+## 1.1.8
+
+### Patch Changes
+
+- 6cad75197: Upgrade all dependencies
+- Updated dependencies [6cad75197]
+  - @prairielearn/html@3.0.6
+
+## 1.1.7
+
+### Patch Changes
+
+- 10cc07dcc: Upgrade all dependencies
+- Updated dependencies [10cc07dcc]
+  - @prairielearn/html@3.0.5
+
 ## 1.1.6
 
 ### Patch Changes