+++ packages/migrations/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @prairielearn/migrations
 
+## 2.0.7
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+- Updated dependencies [f03853d90]
+  - @prairielearn/named-locks@1.5.4
+  - @prairielearn/postgres@1.7.6
+  - @prairielearn/logger@1.0.6
+  - @prairielearn/error@1.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- 6cad75197: Upgrade all dependencies
+- Updated dependencies [6cad75197]
+  - @prairielearn/named-locks@1.5.3
+  - @prairielearn/postgres@1.7.5
+  - @prairielearn/logger@1.0.5
+  - @prairielearn/error@1.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- 10cc07dcc: Upgrade all dependencies
+- Updated dependencies [10cc07dcc]
+  - @prairielearn/named-locks@1.5.2
+  - @prairielearn/postgres@1.7.4
+  - @prairielearn/logger@1.0.4
+  - @prairielearn/error@1.0.5
+
 ## 2.0.4
 
 ### Patch Changes